requirements.txt
pandas
setuptools
typing-extensions
python-dotenv
pathlib
requests
scrapy
beautifulsoup4
transformers
huggingface-hub
nltk
bs4
torch
pyyaml